third_party/blink/web_tests/virtual/popover-hint-disabled/external/wpt/html/semantics/popovers/popover-types-with-hints.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] manuals do not close popovers
  assert_equals: Error, index 1 (<div popover="hint">Hint</div>) expected false but got true
[FAIL] autos close hints but not manuals
  assert_equals: Error, index 1 (<div popover="hint">Hint</div>) expected false but got true
[FAIL] hint is not closed by pre-existing auto
  assert_equals: Error, index 2 (<div popover="">popover 3</div>) expected false but got true
[FAIL] You can nest hint popovers
  assert_equals: Error, index 1 (<div popover="hint">Nested hint</div>) expected false but got true
[FAIL] If a popover=auto is shown, it should hide any open popover=hint, including if the popover=hint is an ancestral popover of the popover=auto. (You can't nest a popover=auto inside a popover=hint)
  assert_equals: Error, index 0 (<div popover="hint">Hint\n    </div>) expected false but got true
[FAIL] If you: a) show a popover=auto (call it D), then b) show a descendent popover=hint of D (call it T), then c) hide D, then T should be hidden. (A popover=hint can be nested inside a popover=auto)
  assert_equals: Error, index 1 (<div popover="">Nested Auto</div>) expected false but got true
[FAIL] If you: a) show a popover=auto (call it D), then b) show a non-descendent popover=hint of D (call it T), then c) hide D, then T should be hidden. (Non-nested popover=hint gets hidden when unrelated popover=autos are hidden)
  assert_equals: Error, index 1 (<div popover="hint">Non-Nested hint</div>) expected false but got true
Harness: the test ran to completion.
